Create README.md
The app is very simple, I made it to demonstrate knowledge in GUI development and linking that to a backend. I designed the front end with Tkinter then connected the appropriate labels and buttons to actions in the backend (Done with sqlite3). In the end, this project works well and can store simple book information and gives user options to delete, update or add more book data to database.